Added column-filter functionality
If the column_filter is not specified or is an empty list, then the table is
not modified.  Else the raw_table_list is filtered based on the values in
the column_filter (i.e., column indexes not specified in the filter is removed).

Each element in the column_filter list must be an integer or a dictionary
with at least the key 'col'.

Specifying an integer in the column_filter list makes sure that column
index is kept (first column is index 0 -- python list indexing).

Specifying a dictionary, gives the optional possibility of specifying the
following keys in the dictionary (note: the keys are mutually exclusive and
specifying more than one has undefined behaviour).

    - filter: filters out (removes) the row, if the content inside this
        column doesn't match (exact string matching of the value of this key
        and the content of the cell).  The value may be a list of strings to
        be matched.

    - regex: filters out (removes) the row, if the content inside this
        column doesn't match.  The value of this key is placed directly into
        `re.match(pattern, string)` as the `pattern` and the cell value as
        the `string`.  Note: Currently we assume that a small amount of
        regex's is used, such that we don't have to deal with compiling of
        regex's, but rely on the built in caching to handle it for us.

Example: This example won't filter out any column, but it demonstrates the
three different ways that you may specify a column-filter.  Just try and
make changes to either one of them, and see how either columns or rows will
be filtered from the resulting table.

    ```{.table}
    ---
    caption: "*Bar* table"
    markdown: yes
    column-filter:
        - 0
        - col: 1
          regex: ".*B|[\\d]"
        - col: 2
          filter: ['C', '3']
    ---
    A,B,C
    1,2,3
    ```