Added backports.csv
Updated pantable and pantable2csv to use backports.csv is we are running under
python 2.

This adds the same handling of unicode and API as in python 3.1+ for the builtin csv
module.

This simplified the code a bit, as backports.csv expects unicode and we could
then use `io.StringIO` and `io.open` for both python 2 and 3, removing some
special handling for python 2.